+++ Ekom/Ekom.Umbraco/Ekom.U10/Services/ImportService.cs
@@ -522,7 +522,7 @@ public class ImportService : IImportService
             {
                 allUmbracoCategories.Add(content);
             }
-            
+
 
             var newParent = string.IsNullOrEmpty(importCategory.ParentIdentifier) ? umbracoRootContent : allUmbracoCategories.FirstOrDefault(x => x.GetValue<string>(Configuration.ImportAliasIdentifier) == importCategory.ParentIdentifier);
 
@@ -1071,7 +1071,10 @@ public class ImportService : IImportService
                 return;
             }
 
-            productContent.SetProperty("title", importProduct.Title);
+            if (!importProduct.PreserveExistingValues || create)
+            {
+                productContent.SetProperty("title", importProduct.Title);
+            }
 
             if (importProduct.Slug != null && importProduct.Slug.Any())
             {
@@ -1084,7 +1087,11 @@ public class ImportService : IImportService
 
             if (!importProduct.PreserveExistingValues)
             {
-                productContent.SetProperty("summary", importProduct.Summary);
+                if (productContent.HasProperty("summary"))
+                {
+                    productContent.SetProperty("summary", importProduct.Summary);
+                }
+
                 productContent.SetProperty("description", importProduct.Description);
             }