+++ Ekom/Ekom.Umbraco/Ekom.U17/Services/ImportService.cs
@@ -526,7 +526,7 @@ public class ImportService : IImportService
             {
                 allUmbracoCategories.Add(content);
             }
-            
+
 
             var newParent = string.IsNullOrEmpty(importCategory.ParentIdentifier) ? umbracoRootContent : allUmbracoCategories.FirstOrDefault(x => x.GetValue<string>(Configuration.ImportAliasIdentifier) == importCategory.ParentIdentifier);
 
@@ -1070,7 +1070,10 @@ public class ImportService : IImportService
                 return;
             }
 
-            productContent.SetProperty("title", importProduct.Title);
+            if (!importProduct.PreserveExistingValues || create)
+            {
+                productContent.SetProperty("title", importProduct.Title);
+            }
 
             if (importProduct.Slug != null && importProduct.Slug.Any())
             {
@@ -1083,7 +1086,10 @@ public class ImportService : IImportService
 
             if (!importProduct.PreserveExistingValues)
             {
-                productContent.SetProperty("summary", importProduct.Summary);
+                if (productContent.HasProperty("summary"))
+                {
+                    productContent.SetProperty("summary", importProduct.Summary);
+                }
                 productContent.SetProperty("description", importProduct.Description);
             }
 
@@ -1118,22 +1124,26 @@ public class ImportService : IImportService
                 }
             }
 
-            if (importProduct.Categories.Count > 1 && allUmbracoCategories != null)
+            if (!importProduct.PreserveExistingValues || create)
             {
-                var categoryIdentifiers = new HashSet<string>(importProduct.Categories.Skip(1));
+                if (importProduct.Categories.Count > 1 && allUmbracoCategories != null)
+                {
+                    var categoryIdentifiers = new HashSet<string>(importProduct.Categories.Skip(1));
 
-                var umbracoCategories = allUmbracoCategories
-                    .Where(x => categoryIdentifiers.Contains(x.GetValue<string>(Configuration.ImportAliasIdentifier) ?? ""))
-                    .ToList();
+                    var umbracoCategories = allUmbracoCategories
+                        .Where(x => categoryIdentifiers.Contains(x.GetValue<string>(Configuration.ImportAliasIdentifier) ?? ""))
+                        .ToList();
 
-                var udis = umbracoCategories.Select(x => x.GetUdi());
+                    var udis = umbracoCategories.Select(x => x.GetUdi());
 
-                var stringUdis = string.Join(",", udis.Select(x => x.ToString()));
+                    var stringUdis = string.Join(",", udis.Select(x => x.ToString()));
 
-                productContent.SetValue("categories", stringUdis);
-            } else
-            {
-                productContent.SetValue("categories", "");
+                    productContent.SetValue("categories", stringUdis);
+                }
+                else
+                {
+                    productContent.SetValue("categories", "");
+                }
             }
 
             if (importProduct.EnableBackorder)